Support ALIGN_WITH_INPUT output section attribute
Previously the linker used to align physical addresses
when users used to specify LMA regions for output sections.

This change makes the linker to work very similiar to GNU linker
when using LMA regions for output sections when assigning physical
addresses.

The change also adds an important keyword ALIGN_WITH_INPUT that allows
the difference in VMA to be captured and made intact for output sections
that specify this attribute.

The documentation is updated with a new section that talks about
controlling physical addresses using linker scripts.

The patch is accompanied by many tests that check for functionality.

Resolves #168
Resolves #56
Resolves #53

Signed-off-by: Shankar Easwaran <seaswara@quicinc.com>