Set emulation mode based on input files
Improve linker to set emulation mode based on input files passed
to the linker.

Linker is enhanced to read the ELF header if its an ELF file or
interpret the target triple from LLVM bitcode files.

This allows eld to be integrated with the build system and does not
require extra flags.

The linker tries to figure out the emulation only if the target
is not known to the linker(If the user does not use the sym link
wrappers).

There is no functionality change if the wrappers are used.

Resolves #279

Signed-off-by: Shankar Easwaran <seaswara@quicinc.com>