[ARM][AArch64] Fix computation of local-exec TLS relocations
This commit fixes computation of local-exec TLS relocations for
ARM and AArch64. Previously, we were not using the correct start offset
in static thread local storage block for computing the relocation.

Lets understand this with the help of an example:

             <--------- Static thread local storage block  ------------------>
| --------- | ----------------- | --------------------- | ------------------ |
     TCB    TP  Reserved space      Alignment padding      Thread variables

ARM/AArch64 local-exec relocation resolves to:
  SymbolValueInTLSSegment + Reserved Space + Alignment padding

Previously, we were not accounting for the alignment padding. But why is
this alignment padding required? It is required so that the symbol
alignments assumed at the compile/link time is satisfied at runtime as
well. The exact condition that should be satisfied is:

  VMA(TP) + Reserved space + Alignment padding congruent to TLSSegment[vaddr] (modulo TLSSegment[align])

This condition becomes more clear with the help of an example:

  StartVMAForStaticThreadVariables = VMA(TP) + Reserved space + Alignment padding

  StartVMAForStaticThreadVariables should be congruent to TLSSegment[vaddr] (modulo TLSSegment[align])

Resolves #302

Signed-off-by: Parth Arora <partaror@qti.qualcomm.com>