Fix intermittent exit code 141 seen with DefaultImageBase.test
This commit fixes intermittent exit code 141 seen with
DefaultImageBase.test. The error code 141 indicates that the program
has exited due to SIGPIPE signal. It happens when a command (writer) is
still writing to the pipe, but there reader is no reader. The root cause
here was that 'grep LOAD -m 1' exits immediately after finding a match.
This causes SIGPIPE in the readelf command that was passing output to
the grep.

Signed-off-by: Parth Arora <partaror@qti.qualcomm.com>