Fix fragment padding size computation when output section is unaligned
This commit fixes fragment padding size computation when output section
address is unaligned. The fragment padding size computation
(incorrectly) assumed that the fragment offset is congruent to the
output section address (modulo fragment alignment). This assumption is
violated when the output section address is explicitly specified in the
linker script.

This change is causing UnalignedCrash.test to fail because the test
expects the link to fail but now the link is passing. The link passing
seems correct as both bfd/lld pass as well for this test case.

Resolves #346

Change-Id: I983658470b03583774593c4889a26dd6c7fff800
Signed-off-by: Parth Arora <partaror@qti.qualcomm.com>