Use ELD_TARGETS to create test configs instead of LLVM_TARGETS
This commit changes for which targets the lit test configs are created.
We are using LLVM_TARGETS_TO_BUILD to determine the target list for
testing. We should instead use ELD_TARGETS_TO_BUILD because it is not
necessary that eld supports all the targets that the LLVM does. This
commit makes the change to use ELD_TARGETS_TO_BUILD for creating lit
test configs.

Signed-off-by: Parth Arora <partaror@qti.qualcomm.com>