[RISCV][Xqci] Remove c.lui absolute relaxations
These are not valid, as we do not verify that the page-alignment that we
checked for during relaxation will be preserved in the final object.
Without these checks, we may end up creating a `c.lui` which cannot
materialize the intended address, because the address is no longer
aligned due to relaxation.

Fixes #455

Signed-off-by: Sam Elliott <aelliott@qti.qualcomm.com>